feat: move invoices and receipts section to admin/sections
